refactor: move directory of HotTrend DiscountItems
