+++ apps/web/src/app/(home)/[category]/HotTrend.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 
-import { Convenience } from '../type';
+import { Convenience } from '../../type';
 
 interface HotTrendProps {
   convenience: Convenience;
